requirements.txt
kubernetes==10.0.1
pymongo==3.10.1
python-dotenv==0.10.5